[refactor][xs]: moved getToken to commons and updated tests
