+++ lib/push.js
@@ -3,7 +3,7 @@ const creds = require('./config')
 const crypto = require('crypto')
 const fs = require('fs')
 const { logger } = require('./utils/log-handler')
-const { checkDpIsThere } = require('./utils/common')
+const { checkDpIsThere, getToken } = require('./utils/common')
 const { spinner } = require('./utils/tools')
 const path = require('path')
 const request = require('request-promise-native')
@@ -44,27 +44,6 @@ const push = async() => {
   const message = '🙌  your Data Package is published!\n'
   const url = '🔗  ' + urljoin(config.server, config.username, dpjson.name)
   logger(message + url, 'success', false, spinner)
-}
-
-const getToken = async(config) => {
-  spinner.text = 'Authenticating...'
-  let res =  await axios.post(
-    urljoin(config.server,'/api/auth/token'),
-      {
-        'username': config.username,
-        'secret': config.secretToken
-      }
-    ).catch(err => {
-      if (err.code === 'ECONNREFUSED' || err.code === 'ENOTFOUND') {
-        logger(`Not able to connect to ${config.server}`, 'error', true, spinner)
-      }
-      const statusCodes = [400, 404, 403, 500]
-      if (err.response && statusCodes.indexOf(err.response.status) > -1) {
-        logger(err.response.data.message, 'error', true, spinner)
-      }
-      logger(err.message, 'error', true, spinner)
-    })
-  return res.data.token
 }
 
 const getFileData = async (config, fileInfo, token) => {
@@ -180,5 +159,4 @@ module.exports.getFileData = getFileData
 module.exports.getFileInfo = getFileInfo
 module.exports.getFileList = getFileList
 module.exports.getFilesForRequest = getFilesForRequest
-module.exports.getToken = getToken
 module.exports.uploadFile = uploadFile
