+++ lib/utils/common.js
@@ -1,6 +1,8 @@
 const axios = require('axios')
 const config = require('../config')
 const fs = require('fs')
+const urljoin = require('url-join')
+const {spinner} = require('./tools')
 
 const parseIdentifier = (dhpkgid) => {
   const [publisher, pkg, path ] = dhpkgid.split('/')
@@ -38,7 +40,30 @@ const getMetadata = async (publisher, pkg, sUrl=getServerUrl() ) => {
   return res.data
 }
 
+
+const getToken = async(config) => {
+  spinner.text = 'Authenticating...'
+  let res =  await axios.post(
+    urljoin(config.server,'/api/auth/token'),
+      {
+        'username': config.username,
+        'secret': config.secretToken
+      }
+    ).catch(err => {
+      if (err.code === 'ECONNREFUSED' || err.code === 'ENOTFOUND') {
+        logger(`Not able to connect to ${config.server}`, 'error', true, spinner)
+      }
+      const statusCodes = [400, 404, 403, 500]
+      if (err.response && statusCodes.indexOf(err.response.status) > -1) {
+        logger(err.response.data.message, 'error', true, spinner)
+      }
+      logger(err.message, 'error', true, spinner)
+    })
+  return res.data.token
+}
+
 module.exports.getServerUrl = getServerUrl
 module.exports.checkDpIsThere = checkDpIsThere
 module.exports.parseIdentifier = parseIdentifier
 module.exports.getMetadata = getMetadata
+module.exports.getToken = getToken
